Move Toast code into its own ES6 module.
