+++ html/dashboard.html
@@ -104,6 +104,7 @@
     <!-- Dynamically inserted -->
     <footer></footer>
 
-    <script src="js/dashboard.js"></script>
+    <script type="module" src="js/toasts.mjs"></script>
+    <script type="module" src="js/dashboard.js"></script>
 </body>
 </html>
